+++ broadcast-chan/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Fixed Haddock links.
 * Generalised `readBChan`/`writeBChan` in `BroadcastChan.Throw` to use
   `MonadIO`.
+* Add non-blocking `tryReadBChan` in `BroadcastChan` and `BroadcastChan.Throw`,
+  fixes [#12](https://github.com/merijn/broadcast-chan/issues/12).
 
 0.2.1.2 [2022.08.24]
 --------------------